Enable no-unused-expressions and a11y/anchor-is-valid
Summary:
This diff renables no-unused-expressions.  This time we'll wait
until all of the lint has been fixed in webapp/master before
landing this diff onto khan-linter/master.  It also adds anchor-is-valid.
I decided to add these rules together in order to cut down on the
number of times we have to updated webapp/master with changes to
keep the linter happy with these new additions.

Test Plan:
- run ka-lint javascript in webapp, see that it outputs 174 unique files for a11y/anchor-is-valid
- supress anchor-is-valid erros using update_eslint_disable_lines.js
- run testing.lint_test.Lint_templates.test see that page_with_js_errors.html failes
- suppress no-unused-expressions error

Reviewers: benkraft, scott

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D43805