Ignore SIGILL from eslint in addition to SIGSEGV
I just saw a similar eslint crash to the segfaults I had been seeing,
but this time with SIGILL.  I'm guessing it's the same issue -- it's
somehow trying to execute some random memory location, and sometimes it
happens to get lucky and find one that it has permission to execute.  So
we'll retry on that too.

Test plan: cross fingers a lot

Reviewers: csilvers