Enable the aria-role rule to check for valid role properties
Summary:
I've been working with modals and ARIA tags lately. There's a lot of risk around providing invalid ARIA tags that look fine but really aren't. The linter can help us catch these.

Just a note on the parameter--we have some uses of `role` in our components. It's possible that these could be exposed if we passed all of the props through to a DOM element. I'll reach out to the owners and see if we can remove this DOM-specific rule in the future.

Test Plan: * Run `ka-lint javascript` and confirm that there are no errors.

Reviewers: benkraft

Subscribers: