Refactor linting to use python logger
Summary:
Addresses Jira ticket: https://khanacademy.atlassian.net/browse/INFRA-1262

To log output, Khan linters have previously relied on a slightly modified implementation of six.print_ rather than using the Python logging library. In considering how to log linter output to a logfile for profiling, the current way was getting messy. Instead, we refactor to using the Python logging library where we can to easily allow us to log to two different places, both regular stdout/stderr (for INFO or higher -- unless in verbose mode in which case we do DEBUG or higher) in addition to the new lint logfile (for DEBUG or higher).

The actual printing of lint errors by line (with formatting to point to error) remains printed by six.print as it does not print to the console properly using logging. There is a TODO included to further investigate if there is a way around this, so as to get rid of our custom six.print_ entirely.

Planned follow-up diffs include:
- Uploading logfile to GCS
- Printing timing for individual linters in webapp's runlint.py

Test Plan:
1. Run `ka-lint` (without verbose) and observe verbose output in a new logfile called  `lint_logfile.log`
2. Run `ka-lint --verbose` and observe verbose output in both the logfile and in stdout/stderr
3. Test the above steps in webapp as well (while introducing some poor formatting in various python and js files)

Reviewers: benkraft, csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D45987