When we get a UnicodeDecodeError, do not print the entire file contents.
`why.args[1]` is really useful for IOError: it is a human-readable
description of what went wrong.

`why.args[1]` is really *not* useful for UnicodeDecodeError: it is the
contents of the file that we couldn't decode.  This could be huge.

So go back to printing a useful message for those.

Auditors: jacqueline

Test plan:
In Khan/internal-services I ran
   python3 ~/khan/devtools/khan-linter/runlint.py --blacklist=no buildmaster/image/buildmaster/models.py
and it said:
   SKIPPING lint of models.py: 'ascii' codec can't decode byte 0xe2 in position 14235: ordinal not in range(128)