Add support for batched filepaths
Summary:
This commit corresponds to the arcanist linter change - see https://phabricator.khanacademy.org/D46540.

Prior to this commit, our ka-lint script would support several filepaths as args if we wanted to run the linter on a set of filepaths, but the way arcanist will send these to ka-lint is as a string of space-separated filepaths (e.g. "/path1 /path2 /path3") which would be read as a single arg. This commit adds a checks for this and parses accordingly.

Test Plan:
Run ka-lint in the following ways:
1. `ka-lint` (with no args)
2. `ka-lint runlint.py linters.py` (with two args in the manner previously supported)
3. `ka-lint "runlint.py linters.py"` (with a single arg containing several filepaths, not previously supported)
4. Verify that in each case, both filepaths are being linted (or all filepaths in the case of #1)

Reviewers: csilvers

Subscribers: