Fix the extra-linter to do a better job sharing its stderr on error.
Before, we assumed that if the extra-linter had an error linting, it
would emit its error to stderr and emit nothing to stdout.

But that's not true for us anymore, because our extra-linter runs
`make deps`, which emits output to stdout.

So now I use a more exact check: not just whether stdout is empty or
not, but whether it has any content in it that could plausibly be a
lint-line.

Auditors: jacqueline, benkraft

Test plan:
I ran
   make linc
on a commit that had an import-error when running the extra linters,
and saw the exception printed, whereas before it would just fail
silently.