Account for line num offset on ktlint blank lines
ktlint reports that needless blank lines at the end of the file occur
one line number later than they actually do, if there is only one blank
line. In other words, the line number of the lint error is one higher
than the last line in the file.  This causes an index error when we try
to process this error message in the linter.

Test Plan:
- add just one extra blank line to a file and lint it!

Auditors: colin