Upgrade ESLint to v5
Upgrade the main ESLint to v5. Upgrade all eslint pugins to latest
(including `babel-eslint`).

Fixed up eslint rules that are now deprecated or irrelevant due to our
usage of Prettier (`flowtype/boolean-style`, `no-unexpected-multiline`).

Sets `ecmaVersion` to 2018 as the experimental-object-rest-spread rule
is deprecated and support for object rest spread is introduced in
`ecmaVersion=2018`.

`ecmaFeatures` has been deprecated so that's also removed.

Test Plan:
I connected this to our mobile repo on a branch prepping for an upgrade
to React Native 0.57 (includes a Babel 7, which is why this revision is
needed) and ran `arc lint --everything` without errors.