Use a better way of telling whether a logger does debug-logging.
This is confusing: in python logging, there are two types of
log-levels.  One is on the logger itself, and one is on each of the
logger's handlers.  For a log-message to get emitted via a particular
handler, its level must be higher than *both* log-levels: the level on
the logger itself and the level on the relevant handler.

Before we were just looking at the handler, which led us to make the
wrong is-verbose determination for the default logger (which has an
INFO level on the logger and an EVERYTHING level on the handler).  I
fixed it to look at both, like it ought.

While investigating this, I realized that `_is_vebose()` really cares
about whether we're logging to //stout//.  So I added that check as
well.

Auditors: jacqueline

Test plan:
To test the default logger, I ran
   tools/runtests.py testing.lint_test.Lint_content_render.test
in webapp, where there was a lint error there, and saw the error
printed, whereas it wasn't before.

To test the khan-linter logger, I ran
   ka-lint content_render
in webapp and saw the same error printed.

To test `_is_verbose()` being True, I ran
   ka-lint -v content_render
in webapp and saw the same error printed (along with a ton of other
text too :-) ).