Merge `CKEDITOR_BASEPATH` into `ckeditor.js` package file
We have a teeny weeny file called `ckeditor-path-fix.js` that just specifies the `CKEDITOR_BASEPATH`, which is breaking the js file dependency test,. It's so small athat it really shouldn't be it's own file anyway, so let's just merge it in and be done with it.

The merging of the file's contents happened in a previous commit: https://phabricator.khanacademy.org/rCD46d98a9b4d70

Test Plan:
Run the following tests:
- `tools/runtests.py build.js_css_packages.unused_js_files_test.UnusedJsFilesTest.test_unused_files`
- `tools/runtests.py --max-size=huge build/js_css_packages/third_party_deps_sufficient_test.py`
and verify that there are no errors related to `ckeditor.js`. (There are currently other unrelated errors.)

Then login as an admin,
navigate to a content item in the CMS (such as `/devadmin/content?mp-r-id=3Z_Lng8=#x00000000-x905db83d-x4b0eb531-x35d0aeea-Topic:xd86de4c8`),
and verify that the editing functionality is working correctly.

Auditors: #webpack