Add eslint-plugin-import
Summary:
This eslint plugin has some useful rules.  In particular we'd like to
make use of the "extensions" rule to ensure that all imports have
extensions.  In a recent code review we spotted an import that was
missing its extension.  I'll add this rule to .eslintrc in webapp
in a separate commit.

Test Plan: n/a

Reviewers: #fe-infra