Fix linter message line split bug
Summary: The linter splits the stdout lines on ':'. However, if the failure message contains ':', such as 'Unexpected spacing before ":"', this throws a RuntimeError instead of displaying the lint failure. Fixing the linter to only split into a maximum of 4 chunks.

Test Plan: - ran arc lint against a file with the 'Unexpected spacing before ":"' error before and after the fix

Reviewers: colin, jacqueline, csilvers

Reviewed By: jacqueline, csilvers

Subscribers: csilvers, kevinb

Differential Revision: https://phabricator.khanacademy.org/D55734