Use flow's no-unused-expressions instead of the base eslint one
Summary:

We currently use ESlint's built-in `no-unused-expressions` rule. In our Flow code, this leads to false positives because the rule sometimes doesn't understand Flow-isms (exhaustive type assertions such as `(myEnum: empty)` in the `default` case of a `switch`). There is a Flow `no-unused-expressions` which is based on the core one which is a bit smarter. Let's use it in Flow contexts.

Ref: https://github.com/eslint/eslint/issues/9877

Test Plan:

Check out this diff and then run `arc lint` against a Flow codebase.

Reviewers: benkraft, kevinb, john, #fe-infra

Subscribers: