Revert "Upgrade ktlint to 0.34.1"
This reverts commit 7944b4e54ddc0b50f952dd521109df012c7b1f89.

It's not clear how they got in -- perhaps more code being merged between
testing the lint-fix commit and deploying the ktlint change -- but there
are some failures on webapp master, so we'll need to fix those up.  In
the meantime, we're reverting khan-linter so we can do other deploys.

Test plan: crossed fingers

Auditors: jeremy, #devops