Upgrade ktlint to version 0.34.2
Summary:

This is attempt #2 upgrading `ktlint` to the latest (v0.34.2). There will be an accompanying diff for `webapp` that fixes all the new linting issues (mostly enforcement of spacing around braces, etc).

Test Plan:

1. Switch to this diff
2. Run `ka-lint` on the `webapp` repo

If you don't have the webapp diff applied, there'll be a bunch of lint warnings. If you do, there should be none!

Reviewers: aasmund, benkraft

Subscribers: #infrastructure