Add basic support for go linting in khan-linter.
Summary:
For now, the only "linting" we do is to make sure `gofmt` doesn't have
any formatting cleanup suggestions.  If it does, we emit them.
`gofmt` emits diffs, not lint errors, so it's not a great fit for
khan-linter right now.  We may try to clean that up somehow but at
least now we can detect gofmt errors via the normal linting framework.

It would be nice to also run `go vet`, but that can't run file-by-file, so is
an even worse fit for khan-linter.  It would also be nice to run `go lint`
but that has false positives with no way to turn them off, so we'll have
to wait for a final decision on what linting library to use for go, first.

Issue: none

Test Plan:
In services/mobile-data/zero_rating_test.go I added a `;` at the end of
   khantest.Run(t, new(zeroRatingSuite));
I then ran
   ka-lint services/mobile-data
and it emitted this:
```
services/mobile-data/zero_rating_test.go:1: F1 FORMAT ERROR(S): run `ka-lint services/mobile-data/zero_rating_test.go` for details
@@ -139,5 +139,5 @@
 }

 func TestZeroRating(t *testing.T) {
-       khantest.Run(t, new(zeroRatingSuite));
+       khantest.Run(t, new(zeroRatingSuite))
 }
```
I then ran `arc lint` and it said:
```

   >>> Lint for services/mobile-data/zero_rating_test.go:

   Error  (S&RX) F1
    FORMAT ERROR(S): run `ka-lint services/mobile-data/zero_rating_test.go`
    for details

    >>>        1 package mobile_data
                ^
```
which is not ideal, but at least it's catching the error.

Reviewers: benkraft, Kai, jacqueline

Reviewed By: benkraft, jacqueline

Subscribers: kevinb, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D58169