Fix up go linting in several ways
This commit fixes a few problems with how we were calling out to
golangci-lint.

First, we fix up how we build and run the binary.  In particular,
`khan-linter` is a tool that's typically used to run on another
repository.  It has its own deps -- in this case in the
form of a go module -- which the calling repo need not import.  So we
want to make sure we build `golangci-lint` in the `khan-linter` module,
and then run the built binary in the caller's go module.  (As it
happens, our go module uses vendoring, but that's not actually at issue
here, except in the precise syntax of the `go build` we need.)  Sadly,
`go run` can't do that, so we need to do `go build` and then call out to
the built binary.  (Luckily, `go build` has caching too so this is no
slower.)

Second, we fix up how we call out to `golangci-lint`.  The way
`khan-linter` likes to do things is to pass a list of files to the
linter.  While `golangci-lint` claims to support such operation, it's
actually very limited: the files must all be in a single directory.
(Older versions seem to have in some cases silently dropped the error;
newer versions report it correctly -- so this will be landed along with
a separate upgrade to v1.21.0.)  So, we do what it wants: we pass in a
list of directories we want it to lint (which it sees as packages and
handles correctly) and then filter the results for the files we actually
care about.  This probably doesn't have much of a perf cost: most static
analysis in Go happens at the package level anyway.

Finally, we fix up error handling: we complain if `golangci-lint` output
any stderr or exited nonzero.  This should make it easier to catch such
problems in the future.

In a separate diff to be deployed after this one, I'll update webapp's
Makefiles to simplify them now that this all works right.

This will need to be deployed after D58739 and D58744, to avoid
lint errors on master.

Test Plan:
`git ls-files '**.go' | xargs ka-lint` passes on webapp master.  If I
introduce a lint error, it fails; and in neither case does it modify
webapp's `go.mod`.

Reviewers: csilvers, Kai, davidbraley

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D58732