Add @khanacademy/eslint-plugin
Summary:
This diff adds our own eslint plugin with our own custom rules.
We'd like to start using the imports-requiring-flow rule right
away.  It allows us to list particular npm packages or source
modules that require an // @flow in the file requiring them.

Test Plan:
See https://github.com/Khan/eslint-plugin-khan/pull/2 for
the test plan.  The only difference is that we need to use
"@khanacademy" for the plugin name instead of "khan" in the configuration.

Reviewers: #fe-infra