Add a dependency on graphql-schema-linter.
I'm adding a new linter!, and this is the first step.  I'm doing it as
an audit so the commit for the actual change isn't too big to view.
(And also because this part is not interesting.

To create this, I just ran
   npm install

Issue: https://khanacademy.atlassian.net/browse/INFRA-4339

Auditors: davidbraley, kai, benkraft

Test plan:
I ran
   node_modules/.bin/graphql-schema-linter ../../webapp/services/content/schema.graphql
and it ran my new code -- I could tell because it gave an error in a
file I had added!