Add a linter for sdl (graphql "schema definition language") files.
Summary:
Mostly it's to make sure we document all our schema fields.

We use the graphql-schema-linter tool to do our work.  We just have to
parse the output and deal with it.

As with eslintrc, we define the linter here, but allow the repo to have its
own config file.  For now, no repo does and so they all use the default
config file I define here, which is easy to satisfy.  That way, I can
deploy this before fixing up webapp to satisfy the "real" rules we want,
which we'll put into webapp's config.

Issue: https://khanacademy.atlassian.net/browse/INFRA-4339

Test Plan:
I ran
   ka-lint ../../webapp/services/assignments/schema.graphql
and it ran and gave no error.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: dhruv, davidbraley, Kai, kevinb

Differential Revision: https://phabricator.khanacademy.org/D61552