Improve schema auto-fixing, and fix a bug in our arc auto-fix util.
I didn't realize, when I first wrote the arc auto-fixers for our new
schema linter, that our utility function really wanted the
text-to-delete and the text-to-add to be on one line, namely the line
with the reported lint error.  Nor did I realize that arc does *not*
have this constraint, and is happy to modify text across lines.

But I wanted to use our existing infrastructure, and it turns out it
*does* have support for "remove text from the end of this line and the
beginning of next", if the text-to-delete has a newline in it.  But
that support was buggy: unlike all the other branches in that routine
it did not do a 'search forward (or backward) to find text-to-delete',
but instead required it to exist at the column reported by the
linter.  It was also unduly complicated.

So I fixe up that utility function, and changed the linter to use it,
which sadly added some complexity.  Hopefully it's comprehensible
complexity, though!

Issue: https://khanacademy.atlassian.net/browse/INFRA-4339

Test plan:
I created a file like this:
```
"""The top-level query object for the content-library service."""
type Query {
    """
    Query for the grades and classes necessary for the class picker UI.
    """
    classPickerForTests(language: String!, country: String!): ClassPicker!
}
```
and ran `arc lint` on it, and it properly added newlines for the first
docstring and removed them for the second.