Don't worry about 80 columns when auto-fixing schema files.
Maybe we should do something about having docstrings fit in 80 cols,
but for now we're not going to worry about it.  That way, we can do
autofixes that make the line longer, e.g. by putting the first line of
the docstring on the same line as the leading triple-quote.

Auditors: benkraft

Test plan:
I ran
   arc lint services/*/*.graphql
in webapp and before this change and after, and after the change it
auto-fixed a few "add blank lines before this field" that it had
refused to auto-fix before.