Go to a two-pronged approach for defining external schema types.
Summary:
Before this diff, when a schema referenced a type defined in a
different schema file, we just made up a "fake" definition of that
type.  That is actually fine for most types, but not for interfaces,
because we have a lint rule that depends on whether a field is defined
in the interface or not, and thus we need the real interface
definition.

I resolve that by loading in the "real" type from other .graphql
files, but *only* for files in the same directory.  That avoids a
separate problem with "duplicate definition" caused by `@external`.
Since folks are unlikely to implement interfaces defined on a
different service, and all graphql files in a single service are in
the same dir by style, that should work ok in practice.

Test Plan:
I ran
   ka-lint services/*/*.graphql
without it crashing.  (It still has a ton of lint errors I need to
resolve.)

I then ran
   ka-lint services/content/article.graphql
and it gave no errors, whereas before this change it complained about
missing comments on a bunch of interface fields, thus showing it was
properly including the "real" interface type.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: jvoll, kevinb

Differential Revision: https://phabricator.khanacademy.org/D61790