Handle utf-8 / unicode properly
I thought `open(..., 'U')` meant we read files as unicode, but it's
not true!  (`U` is for windows line endings, and is deprecated/removed
in both python2 and python3, so I removed it as part of this diff.)

Instead, I discovered python3 automatically converts to unicode and
python2 doesn't.  I've refactored the code to be consistent on both of
them.

It turns out that we can't always interpret the input to utf-8,
though; we use the git linter even on binary files (such as jpeg
files).  Perhaps we shouldn't, but we do.  So I now mark that linter
as "do not try to convert to unicode", and changed the linter to work
properly on binary blobs in python3.

Test plan:
I ran
   webapp% ka-lint services/*/*.graphql
   internal-services% python3 ~/khan/devtools/khan-linter/runlint.py --blacklist=no buildmaster/image/buildmaster/*
without error.