Use a dummy field when declaring dummy types.
I was using `id`, but if we have `extend type Foo` -- necessitating
that we add a dummy type Foo -- then we run the risk that the extender
is defining a field named `id` and we get a collision.  Avoid
collisions by using a dummy name that nobody is likely to use in real
life.

Auditors: benkraft

Test plan:
`ka-lint services/*/*.graphql` in webapp