Use local eslint executable if available
Summary:
This diff is intended to make a little progress on ADR #273.

In D62269 I updated webapp to include all necessary eslint plugins
and verified that the local eslint executable can be run with the
same results as ka-lint which until this diff uses the eslint executable
in this repo.

This diff updates khan-linter to use the local eslint executable in
the repo of the file being linted if it exists.  If it doesn't it falls
back to the eslint executable in the khan-linter repo.

Issue: none

Test Plan:
- in webapp (has its own eslint): ka-lint javascript/about-package
- in kpi-dashboard (no local eslint): ka-lint src/team

Reviewers: #frontend-infrastructure, benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D62270