Defer to the repo defining a yaml file, to say how to lint it.
This is the beginning of the process of moving all the linters
from being owned by khan-linter to being owned by the individual repos.
Now yaml files are entirely linted by code in webapp/, and we just shell
out to it (as a provisional measure while the migration is happening).

Note this means that only webapp can lint yaml files, but it's the only
one that's trying to do so currently, so that's not actually a change in
behavior.

Issue: TODO

Test plan:
I ran
    echo 'a: "b"%' > /tmp/bad.yaml
    ka-lint /tmp/bad.yaml
and saw it say:
    SKIPPING lint of ../../../../../var/tmp/bad.yaml: no yaml linter binary defined in this repo

I then moved bad.yaml into webapp and ran it again, and TODO