Allow lesshint to report errors past the end of the file.
Apparently there's a whole class of errors that lesshint finds only
after whole-file analysis, that it deicdes to report as having after
the last line of the file.  This was causing trouble when we looked at
that non-existent line for `@Nolint` directives.  Now we don't.

Auditors: benkraft

Test plan:
Both of these now complete without crashing:
   ka-lint stylesheets/settings-package/toggle-switch.less
   ka-lint stylesheets