Allowing turning off linting-on-push via an envvar.
Summary:
A big complaint we've heard about the dev workflow is how long git
operations take.  This is in large part due to the hooks we have,
including the hook that lints on push.  (Since linting can be very
slow.)

This commit allows people to turn off that hook, to match github's
behavior, which does such linting asynchronously -- in this case it
will be done by the jenkins test that runs at push-time, or failing
that at deploy time.

If you want the new behavior, you can get it by setting
   GIT_LINT=no
in your .bashrc or similar.

The default is to keep the old behavior, based on the feedback at
   https://khanacademy.slack.com/archives/C02NMB1R5/p1587710103071400

Issue: https://khanacademy.slack.com/archives/C8Y4Q1E0J/p1587600631272500

Test Plan:
   ./test_githook_pre_push.sh

Reviewers: avidal, davidbraley, benkraft

Reviewed By: davidbraley, benkraft

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D62596