Add support for the `--stdin`/etc flags, to match webapp's ka-lint.
Summary:
As part of the move of ka-lint from khan-linter to webapp, I'd like to
make the two drop-in replace-able with each other.  That means they
need to support the same flags -- or, at least, that this old version
supports all the flags that the new version does.

This commit adds that.  I add `--stdin` for real, and I add `--fix`
and `--fast` as noops.

Test Plan:
In webapp, I ran
   ka-lint -v main.py
   ka-lint -v --fast --fix main.py
   echo main.py | ka-lint -v --stdin
and they all succeeded with success, having linted one file.

I then ran
   ka-lint -v
   ka-lint -v --stdin </dev/null
and they both exited without doing any work.

Reviewers: benkraft, davidbraley

Reviewed By: benkraft, davidbraley

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D63417