Properly do the network-liveness testing for khan-linter updating.
I finally figured out why the `git ls-remote` was timing out so often!
I had intended for it to run in the khan-linter repo, which we know is
rather small.  But it was actually running it whatever directory you
happened to be in (which not be in inside a git repo at all!)  Oops.

Issue: none

Auditors: benkraft

Test plan:
Ran
   ka-lint app.yaml
in webapp/ with success.  Before this change I got a timeout error
when I did that.