Make sure we convert our subprocess input to bytes.
Summary:
Python3 cares that the input is bytes rather than strings.  This is
probably right: even for python2 the filenames could conceivably be
unicode and we should encode them before passing to the subshell.

Test Plan:
Fingers crossed, I don't have a good python3 setup.  But I ran:
    python -c 'print "\n".join(["foo"]).encode("utf-8")'
    python3 -c 'print("\n".join(["foo"]).encode("utf-8"))'
and they returned `'foo'` and `b'foo'` respectively, which seems right
to me.

Reviewers: amos, benkraft

Reviewed By: amos

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D63670