Fix a tricky case where we need to lint on merges.
Normally, when we do a merge commit, we don't lint files that were
already on the remote, on the theory they have already been linted.

This works fine, *unless* your local changes include a modification to
(or addition of) a linter.   In that case, some files-on-remote, that
might have been valid when they were pushed, could now be in violation
of the new linter.  So we should lint more aggressively on lint in
that case.

Test plan:
TODO