Fix a tricky case where we need to lint on merges.
Summary:
Normally, when we do a merge commit, we don't lint files that were
already on the remote, on the theory they have already been linted.

This works fine, *unless* your local changes include a modification to
(or addition of) a linter.   In that case, some files-on-remote, that
might have been valid when they were pushed, could now be in violation
of the new linter.  So we should lint more aggressively on lint in
that case.

Issue: https://khanacademy.slack.com/archives/CEA6W0F6F/p1600368820197800

Test Plan:
On my `csilvers` branch, I did:
   echo >> yaml_lint.py
   git commit -am wip
   git fetch origin
   git m origin/master   # brought in 37 changed files
Then I did
   echo "x csilvers x origin/csilvers" |  ~/khan/devtools/khan-linter/githook.py --hook=pre-push x y
and saw it say "LINTING 38 files" whereas before this change it said
"LINTING 1 files".

Reviewers: #devops, benkraft

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D66170